Move white & blacklist back to public settings (oops)
