+++ UI/Web/src/app/admin/manage-metadata-mappings/manage-metadata-mappings.component.html
@@ -2,6 +2,110 @@
 
   <app-loading [loading]="isLoading()" />
   <div ngbAccordion [closeOthers]="false">
+
+    <div ngbAccordionItem="tags-panel" [collapsed]="false">
+      <h2 class="accordion-header" ngbAccordionHeader>
+        <button ngbAccordionButton class="accordion-button" type="button">
+          {{t('tags-header')}}
+        </button>
+      </h2>
+      <div ngbAccordionCollapse>
+        <div ngbAccordionBody>
+          <ng-template>
+            <div class="row mt-4 mb-4">
+              <div class="col-md-6">
+                @if(settingsForm().get('enableGenres'); as formControl) {
+                  <app-setting-switch [title]="t('enable-genres-label')" [subtitle]="t('enable-genres-tooltip')">
+                    <ng-template #switch>
+                      <div class="form-check form-switch float-end">
+                        <input id="enable-genres-status" type="checkbox" class="form-check-input" formControlName="enableGenres">
+                      </div>
+                    </ng-template>
+                  </app-setting-switch>
+                }
+              </div>
+              <div class="col-md-6">
+                @if(settingsForm().get('enableTags'); as formControl) {
+                  <app-setting-switch [title]="t('enable-tags-label')" [subtitle]="t('enable-tags-tooltip')">
+                    <ng-template #switch>
+                      <div class="form-check form-switch float-end">
+                        <input id="enable-tags-status" type="checkbox" class="form-check-input" formControlName="enableTags">
+                      </div>
+                    </ng-template>
+                  </app-setting-switch>
+                }
+              </div>
+            </div>
+
+            <div class="row g-0 align-items-start mb-4">
+              <div class="col">
+                <app-setting-multi-text-field
+                  [title]="t('blacklist-label')"
+                  [tooltip]="t('blacklist-tooltip')"
+                  formControlName="blacklist"
+                />
+              </div>
+            </div>
+
+            <div class="row g-0 mt-4 mb-4">
+              <app-setting-multi-text-field
+                [title]="t('whitelist-label')"
+                [tooltip]="t('whitelist-tooltip')"
+                formControlName="whitelist"
+              />
+            </div>
+
+            @if (licenseService.hasActiveLicense()) {
+              <div class="row g-0 mt-4 mb-4">
+                <app-setting-item [title]="t('tag-weight-label')" [subtitle]="t('tag-weight-tooltip')" [labelId]="'filter-above-weight'">
+                  <ng-template #view>
+                    @if (settingsForm().get('filterAboveWeight')?.value == null) {
+                      {{t('tag-weight-disabled')}}
+                    } @else {
+                      {{settingsForm().get('filterAboveWeight')?.value | tagWeightTitle}}
+                    }
+
+                  </ng-template>
+                  <ng-template #edit>
+                    <select class="form-select" formControlName="filterAboveWeight" [id]="'filter-above-weight'">
+                      <option [ngValue]="null">{{t('tag-weight-disabled')}}</option>
+                      @for(weight of allTagWeights; track weight) {
+                        <option [ngValue]="weight">{{weight | tagWeightTitle}}</option>
+                      }
+                    </select>
+                  </ng-template>
+                </app-setting-item>
+              </div>
+            }
+
+          </ng-template>
+        </div>
+      </div>
+    </div>
+
+    @if (licenseService.hasActiveLicense()) {
+      <div ngbAccordionItem="external-age-rating-panel" [collapsed]="false">
+        <h2 class="accordion-header" ngbAccordionHeader>
+          <button ngbAccordionButton class="accordion-button" type="button">
+            {{t('external-age-rating-header')}}
+          </button>
+        </h2>
+        <div ngbAccordionCollapse>
+          <div ngbAccordionBody>
+            <ng-template>
+              <p>{{t('external-age-rating-description')}}</p>
+              <app-age-rating-mapper [form]="settingsForm()" controlName="externalAgeRatingMappings"
+                                     [sourceLabel]="t('external-age-rating-label')"
+                                     [addLabel]="t('add-age-rating-mapping-label')"
+                                     [removeLabel]="t('remove-age-rating-mapping-label')"
+                                     [strValues]="MangaBakaAgeRatings"
+              />
+            </ng-template>
+          </div>
+        </div>
+      </div>
+    }
+
     <div ngbAccordionItem="age-rating-panel" [collapsed]="ageRatingCollapsed()">
       <h2 class="accordion-header" ngbAccordionHeader>
         <button ngbAccordionButton class="accordion-button" type="button">
